+++ docs/.vuepress/components/ValidationUrl.vue
@@ -1,0 +1,73 @@
+<template>
+  <div class="form-composition library-example">
+    <div class="col">
+      <dynamic-form
+        :id="testForm.id"
+        :fields="testForm.fields"
+        :options="testForm.options"
+        @change="updateForm"
+      />
+      <p class="hint">
+        Please change the value and press enter (blur) to see the validation
+        working 👆🏻
+      </p>
+    </div>
+    <div class="col">
+      <pre>{{ formData }}</pre>
+    </div>
+  </div>
+</template>
+
+<script>
+import {
+  DynamicForm,
+  FormField,
+  FormOptions,
+  FormValidation,
+  url,
+} from '../../../dist/as-dynamic-forms.common';
+
+export default {
+  name: 'ValidationUrlDemo',
+  components: {
+    DynamicForm,
+  },
+  data: () => ({
+    formData: null,
+    testForm: {
+      id: 'validation-url-demo',
+      fields: [
+        new FormField({
+          type: 'url',
+          label: 'Website',
+          name: 'website',
+          value: 'http://bah',
+          validations: [new FormValidation(url, 'Format of url is incorrect')],
+        }),
+      ],
+      options: new FormOptions({ autoValidate: true }),
+    },
+  }),
+  methods: {
+    updateForm(values) {
+      this.formData = values;
+    },
+  },
+};
+</script>
+
+<style lang="scss">
+@import '../styles/styles.scss';
+
+.form-composition {
+  display: flex;
+  justify-content: space-between;
+}
+.col {
+  width: 45%;
+}
+pre {
+  color: white;
+  font-size: 12px;
+}
+</style>