+++ docs/.vuepress/config/isDeployPreview.js
@@ -1,0 +1,1 @@
+module.exports = process.env.hasOwnProperty('DEPLOY_PREVIEW');