+++ docs/.vuepress/data/features.js
@@ -1,0 +1,22 @@
+export default [
+  {
+    title: 'Magical 🧞‍♂️✨',
+    text:
+      'Generating forms quickly from a simple data object, a JSON string or the response from an ajax call using only a DynamicForm component, yup, that easily. ',
+  },
+  {
+    title: 'Customizable 🧑‍🎤👩🏻‍🎤',
+    text:
+      'Vue Dynamic Forms is UI agnostic, plain style ready for you to customize, or you can import one of the default themes',
+  },
+  {
+    title: 'Built-in Validation 🦾',
+    text:
+      'No need to add more dependencies for your input validations, Vue Dynamic Forms comes with a pretty solid built-in validations to make your life and your project easier.',
+  },
+  {
+    title: 'Flexible 🤸',
+    text:
+      'This library is aimed to give an awesome developing experience, you need a really custom input type that is not in the library? Create your own and  attach it vía scoped slots',
+  },
+];