+++ src/components/dynamic-input/DynamicInput.vue
@@ -1,13 +1,18 @@
 <template>
-  <div
-    class="dynamic-input form-group"
-    :class="({ 'form-group--error': hasErrors }, `${formControl.customClass || ''}`)"
-  >
+  <div :class="getClasses">
+    <label
+      class="form-label"
+      :for="formControl.name"
+      v-if="formControl.type !== 'checkbox'"
+    >
+      {{ formControl.label }}
+    </label>
     <input-text
       v-if="
         formControl.type === 'text' ||
         formControl.type === 'email' ||
         formControl.type === 'password' ||
+        formControl.type === 'url' ||
         formControl.type === 'number'
       "
       :formControl="formControl"
@@ -41,15 +46,10 @@
       :onFocus="onFocus"
       :onBlur="onBlur"
     />
-    <label
-      class="form-label"
-      :for="formControl.name"
-      v-if="formControl.type !== 'checkbox'"
-    >
-      {{ formControl.label }}
-    </label>
-    <span class="form-bar"></span>
-    <div v-if="hasErrors">
+    <p class="form-hint" v-if="formControl.helper">
+      {{ formControl.helper }}
+    </p>
+    <div v-if="showErrors">
       <p
         v-for="(errorText, $index) in errorMessages"
         :key="`${$index}`"