+++ src/components/input-text/InputText.vue
@@ -1,6 +1,6 @@
 <template>
   <input
-    :id="formControl.name"
+    :id="formControl.id"
     v-model="formControl.value"
     :name="formControl.name"
     class="form-control"