+++ src/components/input-textarea/InputTextarea.vue
@@ -1,11 +1,14 @@
 <template>
   <textarea
-    :id="formControl.name"
+    :id="formControl.id"
     v-model="formControl.value"
     :name="formControl.name"
     class="form-control"
     :cols="formControl.cols"
     :rows="formControl.rows"
+    @change="valueChange()"
+    @focus="onFocus()"
+    @blur="onBlur()"
   ></textarea>
 </template>
 <script src="./InputTextarea.js"></script>