+++ src/core/utils/form-control.model.js
@@ -1,19 +1,26 @@
 export function FormControl({
+  id,
   type = null,
   value = null,
   validations = [],
   label = null,
+  form,
   name = null,
   customClass = null,
   options = [],
   placeholder = null,
+  rows = null,
+  cols = null,
   errors = {},
   disabled = false,
   valid = true,
   touched = false,
   dirty = false,
+  helper,
 }) {
+  this.id = id || `${form}-${name}`;
   this.type = type;
+  this.form = form;
   this.value = value;
   this.validations = validations;
   this.label = label;
@@ -26,9 +33,13 @@ export function FormControl({
   this.valid = valid;
   this.touched = touched;
   this.dirty = dirty;
+  this.rows = rows;
+  this.cols = cols;
+  this.helper = helper;
 }
 
 export function FormField({
+  id,
   type = 'text',
   value = null,
   validations = [],
@@ -39,7 +50,11 @@ export function FormField({
   options = [],
   placeholder = null,
   inline = false,
+  rows = null,
+  cols = null,
+  helper,
 }) {
+  this.id = id;
   this.type = type;
   this.value = value;
   this.validations = validations;
@@ -50,6 +65,9 @@ export function FormField({
   this.options = options;
   this.placeholder = placeholder;
   this.inline = inline;
+  this.rows = rows;
+  this.cols = cols;
+  this.helper = helper;
 }
 
 export function FormValidation(
@@ -60,4 +78,18 @@ export function FormValidation(
   this.text = text;
 }
 
-export default { FormControl, FormField, FormValidation };
+export function FormOptions({
+  customClass = '',
+  method = 'POST',
+  autoValidate = false,
+  netlify = false,
+  netlifyHoneypot = null,
+}) {
+  this.customClass = customClass;
+  this.method = method;
+  this.autoValidate = autoValidate;
+  this.netlify = netlify;
+  this.netlifyHoneypot = netlifyHoneypot;
+}
+
+export default { FormControl, FormField, FormValidation, FormOptions };