+++ CHANGELOG.md
@@ -5,6 +5,49 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.3.1 - 2020-06-17
+
+## Fixed
+
+- Undefined variables in theme ([#40](https://github.com/alvarosaburido/vue-dynamic-forms/issues/40))
+
+## 0.3.0 - 2020-06-17
+
+## Added
+
+- Customizable Default Theme ([#37](https://github.com/alvarosaburido/vue-dynamic-forms/issues/37))
+- Added missing models to docs ([#36](https://github.com/alvarosaburido/vue-dynamic-forms/issues/36))
+- Helper text/hint ([#26](https://github.com/alvarosaburido/vue-dynamic-forms/issues/26))
+- Url regex validation ([#28](https://github.com/alvarosaburido/vue-dynamic-forms/issues/28))
+- Configurable input `id` ([#31](https://github.com/alvarosaburido/vue-dynamic-forms/issues/31))
+
+## 0.2.2 - 2020-05-27
+
+### Fixed
+
+- Added build command to publish command
+
+## 0.2.1 - 2020-05-27
+
+### Fixed
+
+- Text Area methods ([#29](https://github.com/alvarosaburido/vue-dynamic-forms/issues/29))
+
+## 0.2.0 - 2020-05-26
+
+### Added
+
+- VuePress for Better documentation ([#15](https://github.com/alvarosaburido/vue-dynamic-forms/issues/15))
+- Allow Input validation on change and blur ([#21](https://github.com/alvarosaburido/vue-dynamic-forms/issues/21))
+- Custom fields via scoped slots ([#12](https://github.com/alvarosaburido/vue-dynamic-forms/issues/12))
+- Spread Form options and netlify forms support ([#19](https://github.com/alvarosaburido/vue-dynamic-forms/issues/19))
+
+### Fixed
+
+- Fixed missing github link in npm ([#6](https://github.com/alvarosaburido/vue-dynamic-forms/issues/6))
+- Custom classes ([#16](https://github.com/alvarosaburido/vue-dynamic-forms/issues/16))
+- Textarea validation and events not triggering ([#23](https://github.com/alvarosaburido/vue-dynamic-forms/issues/23))
+
 ## 0.1.2 - 2020-05-05
 
 ### Fixed